refatored a code.
