+++ src/main/java/uk/gov/hmcts/reform/civil/service/flowstate/FlowFlag.java
@@ -26,8 +26,6 @@ public enum FlowFlag {
     IS_JO_LIVE_FEED_ACTIVE,
     DEFENDANT_NOC_ONLINE,
     CLAIM_STATE_DURING_NOC,
-    DEFENDANT_ENGLISH_TO_WELSH,
-    CLAIMANT_ENGLISH_TO_WELSH,
     WELSH_ENABLED,
     BILINGUAL_DOCS;
-  }
+}
