+++ src/test/java/uk/gov/hmcts/reform/civil/handler/tasks/CaseEventTaskHandlerTest.java
@@ -661,8 +661,6 @@ class CaseEventTaskHandlerTest {
                                      Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                                      Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                                     Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                                     Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                                      Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
                                      Map.entry(FlowFlag.DEFENDANT_NOC_ONLINE.name(), false),
@@ -677,8 +675,6 @@ class CaseEventTaskHandlerTest {
                                      Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                                      Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                                     Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                                     Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                                      Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
                                      Map.entry(FlowFlag.DEFENDANT_NOC_ONLINE.name(), false),
@@ -695,8 +691,6 @@ class CaseEventTaskHandlerTest {
                                      Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                                      Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                                     Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                                     Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                                      Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
                                      Map.entry(FlowFlag.DEFENDANT_NOC_ONLINE.name(), false),
@@ -711,8 +705,6 @@ class CaseEventTaskHandlerTest {
                                      Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                                      Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                                     Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                                     Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                                      Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                                      Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
                                      Map.entry(FlowFlag.DEFENDANT_NOC_ONLINE.name(), false),
@@ -727,8 +719,6 @@ class CaseEventTaskHandlerTest {
                               Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                               Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                               Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                              Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                              Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                               Map.entry(FlowFlag.MINTI_ENABLED.name(), false),
                               Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                               Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
