+++ src/test/java/uk/gov/hmcts/reform/civil/handler/tasks/PaymentTaskHandlerTest.java
@@ -93,8 +93,6 @@ class PaymentTaskHandlerTest {
                                                          Map.entry(FlowFlag.DASHBOARD_SERVICE_ENABLED.name(), false),
                                                          Map.entry(FlowFlag.CASE_PROGRESSION_ENABLED.name(), false),
                                                          Map.entry(FlowFlag.BULK_CLAIM_ENABLED.name(), false),
-                                                         Map.entry(FlowFlag.CLAIMANT_ENGLISH_TO_WELSH.name(), false),
-                                                         Map.entry(FlowFlag.DEFENDANT_ENGLISH_TO_WELSH.name(), false),
                                                          Map.entry(FlowFlag.JO_ONLINE_LIVE_ENABLED.name(), false),
                                                          Map.entry(FlowFlag.IS_JO_LIVE_FEED_ACTIVE.name(), false),
                                                          Map.entry(FlowFlag.DEFENDANT_NOC_ONLINE.name(), false),
