+++ src/test/java/uk/gov/hmcts/reform/civil/utils/HmcDataUtilsTest.java
@@ -1881,4 +1881,194 @@ class HmcDataUtilsTest {
             assertThat(result).isFalse();
         }
     }
+
+    @Nested
+    class IsClaimantDQDocumentsWelshTests {
+
+        @Test
+        void shouldReturnFalse_whenApplicant1DQIsNull() {
+            // Given
+            CaseData caseData = CaseData.builder()
+                .applicant1DQ(null)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isClaimantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnFalse_whenApplicant1DQLanguageIsNull() {
+            // Given
+            Applicant1DQ applicant1DQ = Applicant1DQ.builder()
+                .applicant1DQLanguage(null)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .applicant1DQ(applicant1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isClaimantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnFalse_whenDocumentsIsEnglish() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.ENGLISH)
+                .build();
+            Applicant1DQ applicant1DQ = Applicant1DQ.builder()
+                .applicant1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .applicant1DQ(applicant1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isClaimantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnTrue_whenDocumentsIsWelsh() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.WELSH)
+                .build();
+            Applicant1DQ applicant1DQ = Applicant1DQ.builder()
+                .applicant1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .applicant1DQ(applicant1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isClaimantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isTrue();
+        }
+
+        @Test
+        void shouldReturnTrue_whenDocumentsIsBoth() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.BOTH)
+                .build();
+            Applicant1DQ applicant1DQ = Applicant1DQ.builder()
+                .applicant1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .applicant1DQ(applicant1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isClaimantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isTrue();
+        }
+    }
+
+    @Nested
+    class IsDefendantDQDocumentsWelshTests {
+
+        @Test
+        void shouldReturnFalse_whenRespondent1DQIsNull() {
+            // Given
+            CaseData caseData = CaseData.builder()
+                .respondent1DQ(null)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isDefendantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnFalse_whenRespondent1DQLanguageIsNull() {
+            // Given
+            Respondent1DQ respondent1DQ = Respondent1DQ.builder()
+                .respondent1DQLanguage(null)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .respondent1DQ(respondent1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isDefendantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnFalse_whenDocumentsIsEnglish() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.ENGLISH)
+                .build();
+            Respondent1DQ respondent1DQ = Respondent1DQ.builder()
+                .respondent1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .respondent1DQ(respondent1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isDefendantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isFalse();
+        }
+
+        @Test
+        void shouldReturnTrue_whenDocumentsIsWelsh() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.WELSH)
+                .build();
+            Respondent1DQ respondent1DQ = Respondent1DQ.builder()
+                .respondent1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .respondent1DQ(respondent1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isDefendantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isTrue();
+        }
+
+        @Test
+        void shouldReturnTrue_whenDocumentsIsBoth() {
+            // Given
+            WelshLanguageRequirements req = WelshLanguageRequirements.builder()
+                .documents(Language.BOTH)
+                .build();
+            Respondent1DQ respondent1DQ = Respondent1DQ.builder()
+                .respondent1DQLanguage(req)
+                .build();
+            CaseData caseData = CaseData.builder()
+                .respondent1DQ(respondent1DQ)
+                .build();
+
+            // When
+            boolean result = HmcDataUtils.isDefendantDQDocumentsWelsh(caseData);
+
+            // Then
+            assertThat(result).isTrue();
+        }
+    }
 }
